Lead flow along the journey
This query creates a funnel using multiple tables (browse, checkout, purchase) showing how many users moved along each journey step.
Using this data it is possible to find out which stages of the journey are less efficient. These steps should be reviewed in detail to find the root cause of the problem or areas of improvement.